Add files via upload
The directory I was in when running the command was the main directory. I didn't get an output because I didn't give an input in to what directory I wanted to enter. The output isn't an error because this command isn't meant to have no arguments so having no arguments results in no output.